refactor: ExcrementDisplay와 IconAndNumberDisplay 컴포넌트의 파일명 변경
