+++ frontend/src/components/journals/ExcrementDisplay.tsx
@@ -1,6 +1,6 @@
 import { Feces } from '@/components/icon/Feces';
 import { Urine } from '@/components/icon/Urine';
-import IconAndNumberDisplay from '@/components/journals/iconAndNumberDisplay';
+import IconAndNumberDisplay from '@/components/journals/IconAndNumberDisplay';
 
 export default function ExcrementDisplay({ fecesCount = 0, urineCount = 0 }: Props) {
     return (
