+++ frontend/src/components/journals/IconAndNumberDisplay.tsx
